[refactor] GraphQLEndpoint: update imports
